+++ src/main/java/pl/overlookhotel/domain/guest/GuestService.java
@@ -8,7 +8,8 @@ import java.util.List;
 
 public class GuestService {
 
-    private final  GuestRepository repository = ObjectPool.getGuestRepository();
+    private final GuestRepository repository = ObjectPool.getGuestRepository();
+
 
     private final static GuestService instance = new GuestService();
 
@@ -43,12 +44,12 @@ public class GuestService {
         this.repository.readAll();
     }
 
-    public void removeGuest(int id) {
+    public void removeGuest(long id) {
         this.repository.remove(id);
     }
 
 
-    public void editGuest(int id, String firstName, String lastName, int age, boolean isMale) {
+    public void editGuest(long id, String firstName, String lastName, int age, boolean isMale) {
 
         Gender gender = Gender.FEMALE;
 
@@ -59,7 +60,7 @@ public class GuestService {
         this.repository.edit(id, firstName, lastName, age, gender);
     }
 
-    public Guest getGuestById(int guestId) {
+    public Guest getGuestById(long guestId) {
 
         return this.repository.getById(guestId);
     }
