+++ src/main/java/pl/overlookhotel/domain/reservation/ReservationService.java
@@ -29,7 +29,7 @@ public class ReservationService {
         return instance;
     }
 
-    public Reservation createNewReservation(LocalDate from, LocalDate to, int roomId, int guestId) throws IllegalArgumentException {
+    public Reservation createNewReservation(LocalDate from, LocalDate to, long roomId, long guestId) throws IllegalArgumentException {
 
         Room room = this.roomService.getRoomById(roomId);
         Guest guest = this.guestService.getGuestById(guestId);
@@ -53,7 +53,7 @@ public class ReservationService {
 
     }
 
-    public void removeReservation(int id){
+    public void removeReservation(long id){
         this.repository.remove(id);
     }
