+++ src/main/java/pl/overlookhotel/ui/gui/reservations/AddNewReservationScene.java
@@ -12,7 +12,6 @@ import pl.overlookhotel.domain.reservation.ReservationService;
 import pl.overlookhotel.domain.reservation.dto.ReservationDTO;
 import pl.overlookhotel.domain.room.RoomService;
 import pl.overlookhotel.domain.room.dto.RoomDTO;
-import pl.overlookhotel.ui.gui.guests.GuestSelectionItem;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -55,7 +54,7 @@ public class AddNewReservationScene {
         List<GuestSelectionItem> guestSelectionItems = new ArrayList<>();
 
         this.guestService.getAllGuestsAsDTO().forEach(dto -> {
-            guestSelectionItems.add(new GuestSelectionItem(dto.getFirstName(), dto.getLastName(), dto.getId()));
+            guestSelectionItems.add(new GuestSelectionItem(dto.getFirstName(), dto.getLastName(), (int)dto.getId()));
         });
 
         Label roomLabel = new Label("Pokój:");
@@ -78,7 +77,7 @@ public class AddNewReservationScene {
             LocalDate from = fromDateField.getValue();
             LocalDate to = toDateField.getValue();
             int guestId = guestField.getValue().getId();
-            int roomId = roomField.getValue().getId();
+            long roomId = roomField.getValue().getId();
 
             try {
                 this.reservationService.createNewReservation(from, to, roomId, guestId);
